refactor(*): more refactoring
